requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.1.0
openai>=1.40.0

# Vector database
chromadb>=0.5.0

# Utilities
tiktoken>=0.7.0
python-dotenv>=1.0.1
pypdf>=4.2.0
